+++ app/use_cases/projects/__init__.py
@@ -3,10 +3,12 @@ from .delete_project import HardDeleteProjectUseCase, SoftDeleteProjectUseCase
 from .dto import (
     CreateProjectCommand,
     ProjectActor,
+    ProjectLeaderStatsListQuery,
     ProjectListQuery,
     UpdateProjectCommand,
 )
 from .get_project import (
+    GetAllProjectLeadersStatsUseCase,
     GetProjectLeaderStatsDetailsUseCase,
     GetProjectLeaderStatsUseCase,
     GetProjectStatsByMaterialsUseCase,
@@ -24,11 +26,13 @@ __all__ = [
     "GetProjectStatsByMaterialsUseCase",
     "GetProjectLeaderStatsUseCase",
     "GetProjectLeaderStatsDetailsUseCase",
+    "GetAllProjectLeadersStatsUseCase",
     "GetProjectStatsUseCase",
     "GetProjectUseCase",
     "HardDeleteProjectUseCase",
     "ListProjectsUseCase",
     "ProjectActor",
+    "ProjectLeaderStatsListQuery",
     "ProjectListQuery",
     "ProjectRepository",
     "SoftDeleteProjectUseCase",